Move parameter column tests to separate file
