+++ CDP4Composition.Tests/Reporting/DataSourceParameterTestFixture.cs
@@ -1,5 +1,5 @@
 ﻿// --------------------------------------------------------------------------------------------------------------------
-// <copyright file="ReportingDataSourceTestFixture.cs" company="RHEA System S.A.">
+// <copyright file="DataSourceParameterTestFixture.cs" company="RHEA System S.A.">
 //    Copyright (c) 2015-2020 RHEA System S.A.
 //
 //    Author: Sam Gerené, Alex Vorobiev, Alexander van Delft, Cozmin Velciu, Adrian Chivu
@@ -40,7 +40,7 @@ namespace CDP4Composition.Tests.Reporting
     using NUnit.Framework;
 
     [TestFixture]
-    public class ReportingDataSourceTestFixture
+    public class DataSourceParameterTestFixture
     {
         private ConcurrentDictionary<CacheKey, Lazy<Thing>> cache;
 
@@ -48,25 +48,16 @@ namespace CDP4Composition.Tests.Reporting
 
         private Category cat1;
         private Category cat2;
-        private Category cat3;
 
         private SimpleQuantityKind parameterType1;
         private SimpleQuantityKind parameterType2;
         private SimpleQuantityKind parameterType3;
 
         private ElementDefinition ed1;
-        private ElementDefinition ed2p;
-        private ElementDefinition ed2n;
-        private ElementDefinition ed3;
+        private ElementDefinition ed2;
 
-        private ElementUsage eu12p1;
-        private ElementUsage eu12p2;
-        private ElementUsage eu12n1;
-        private ElementUsage eu12n2;
-        private ElementUsage eu2p31;
-        private ElementUsage eu2p32;
-        private ElementUsage eu2n31;
-        private ElementUsage eu2n32;
+        private ElementUsage eu1;
+        private ElementUsage eu2;
 
         [DefinedThingShortName("type1")]
         private class TestParameter1 : ReportingDataSourceParameter<Row>
@@ -106,7 +97,9 @@ namespace CDP4Composition.Tests.Reporting
                 Name = "cat1"
             };
 
-            this.cache.TryAdd(new CacheKey(this.cat1.Iid, null), new Lazy<Thing>(() => this.cat1));
+            this.cache.TryAdd(
+                new CacheKey(this.cat1.Iid, null),
+                new Lazy<Thing>(() => this.cat1));
 
             this.cat2 = new Category(Guid.NewGuid(), this.cache, null)
             {
@@ -114,15 +107,9 @@ namespace CDP4Composition.Tests.Reporting
                 Name = "cat2"
             };
 
-            this.cache.TryAdd(new CacheKey(this.cat2.Iid, null), new Lazy<Thing>(() => this.cat2));
-
-            this.cat3 = new Category(Guid.NewGuid(), this.cache, null)
-            {
-                ShortName = "cat3",
-                Name = "cat3"
-            };
-
-            this.cache.TryAdd(new CacheKey(this.cat3.Iid, null), new Lazy<Thing>(() => this.cat3));
+            this.cache.TryAdd(
+                new CacheKey(this.cat2.Iid, null),
+                new Lazy<Thing>(() => this.cat2));
 
             // Parameter types
 
@@ -156,132 +143,50 @@ namespace CDP4Composition.Tests.Reporting
             this.AddParameter(this.ed1, this.parameterType2, "12");
             this.AddParameter(this.ed1, this.parameterType3, "13");
 
-            this.ed2p = new ElementDefinition(Guid.NewGuid(), this.cache, null)
+            this.ed2 = new ElementDefinition(Guid.NewGuid(), this.cache, null)
             {
-                ShortName = "ed2p",
-                Name = "element definition 2p"
+                ShortName = "ed2",
+                Name = "element definition 2"
             };
 
-            this.AddParameter(this.ed2p, this.parameterType1, "21");
-            this.AddParameter(this.ed2p, this.parameterType2, "22");
-            this.AddParameter(this.ed2p, this.parameterType3, "23");
-
-            this.ed2n = new ElementDefinition(Guid.NewGuid(), this.cache, null)
-            {
-                ShortName = "ed2n",
-                Name = "element definition 2n"
-            };
-
-            this.AddParameter(this.ed2n, this.parameterType1, "-21");
-            this.AddParameter(this.ed2n, this.parameterType2, "-22");
-            this.AddParameter(this.ed2n, this.parameterType3, "-23");
-
-            this.ed3 = new ElementDefinition(Guid.NewGuid(), this.cache, null)
-            {
-                ShortName = "ed3",
-                Name = "element definition 3"
-            };
-
-            this.AddParameter(this.ed3, this.parameterType1, "31");
-            this.AddParameter(this.ed3, this.parameterType2, "32");
-            this.AddParameter(this.ed3, this.parameterType3, "33");
+            this.AddParameter(this.ed2, this.parameterType1, "21");
+            this.AddParameter(this.ed2, this.parameterType2, "22");
+            this.AddParameter(this.ed2, this.parameterType3, "23");
 
             // Element Usages
 
-            this.eu12p1 = new ElementUsage(Guid.NewGuid(), this.cache, null)
+            this.eu1 = new ElementUsage(Guid.NewGuid(), this.cache, null)
             {
-                ElementDefinition = this.ed2p,
-                ShortName = "eu12p1",
-                Name = "element usage 1->2p #1"
+                ElementDefinition = this.ed2,
+                ShortName = "eu1",
+                Name = "element usage 1"
             };
 
-            this.eu12p2 = new ElementUsage(Guid.NewGuid(), this.cache, null)
+            this.AddParameterOverride(this.eu1, this.parameterType1, "121");
+            this.AddParameterOverride(this.eu1, this.parameterType2, "122");
+            this.AddParameterOverride(this.eu1, this.parameterType3, "123");
+
+            this.eu2 = new ElementUsage(Guid.NewGuid(), this.cache, null)
             {
-                ElementDefinition = this.ed2p,
-                ShortName = "eu12p2",
-                Name = "element usage 1->2p #2"
+                ElementDefinition = this.ed2,
+                ShortName = "eu2",
+                Name = "element usage 2"
             };
 
-            this.eu12n1 = new ElementUsage(Guid.NewGuid(), this.cache, null)
-            {
-                ElementDefinition = this.ed2n,
-                ShortName = "eu12n1",
-                Name = "element usage 1->2n #1"
-            };
-
-            this.eu12n2 = new ElementUsage(Guid.NewGuid(), this.cache, null)
-            {
-                ElementDefinition = this.ed2n,
-                ShortName = "eu12n2",
-                Name = "element usage 1->2n #2"
-            };
-
-            this.eu2p31 = new ElementUsage(Guid.NewGuid(), this.cache, null)
-            {
-                ElementDefinition = this.ed3,
-                ShortName = "eu2p31",
-                Name = "element usage 2p->3 #1"
-            };
-
-            this.AddParameterOverride(this.eu2p31, this.parameterType1, "231");
-            this.AddParameterOverride(this.eu2p31, this.parameterType2, "232");
-            this.AddParameterOverride(this.eu2p31, this.parameterType3, "233");
-
-            this.eu2p32 = new ElementUsage(Guid.NewGuid(), this.cache, null)
-            {
-                ElementDefinition = this.ed3,
-                ShortName = "eu2p32",
-                Name = "element usage 2p->3 #2"
-            };
-
-            this.eu2n31 = new ElementUsage(Guid.NewGuid(), this.cache, null)
-            {
-                ElementDefinition = this.ed3,
-                ShortName = "eu2n31",
-                Name = "element usage 2n->3 #1"
-            };
-
-            this.eu2n32 = new ElementUsage(Guid.NewGuid(), this.cache, null)
-            {
-                ElementDefinition = this.ed3,
-                ShortName = "eu2n32",
-                Name = "element usage 2n->3 #2"
-            };
+            this.AddParameterOverride(this.eu2, this.parameterType1, "221");
+            this.AddParameterOverride(this.eu2, this.parameterType2, "222");
+            this.AddParameterOverride(this.eu2, this.parameterType3, "223");
 
             // Structure
 
             this.iteration.TopElement = this.ed1;
             this.ed1.Category.Add(this.cat1);
 
-            this.eu12n1.Category.Add(this.cat2);
-            this.ed1.ContainedElement.Add(this.eu12n1);
-            this.ed1.ContainedElement.Add(this.eu12n2);
+            this.eu1.Category.Add(this.cat2);
+            this.ed1.ContainedElement.Add(this.eu1);
 
-            this.eu12p1.Category.Add(this.cat2);
-            this.ed1.ContainedElement.Add(this.eu12p1);
-            this.ed1.ContainedElement.Add(this.eu12p2);
-
-            this.ed2n.ContainedElement.Add(this.eu2n31);
-            this.ed2n.ContainedElement.Add(this.eu2n32);
-
-            this.eu2p31.Category.Add(this.cat3);
-            this.ed2p.ContainedElement.Add(this.eu2p31);
-            this.ed2p.ContainedElement.Add(this.eu2p32);
-
-            // Product tree:
-            // ["cat1"] +> "ed1"
-            // ["cat2"] +-+> "ed1.eu12n1"
-            // [      ] | +--> "ed1.eu12n1.eu2n31"
-            // [      ] | +--> "ed1.eu12n1.eu2n32"
-            // [      ] +-+> "ed1.eu12n2"
-            // [      ] | +--> "ed1.eu12n2.eu2n31"
-            // [      ] | +--> "ed1.eu12n2.eu2n32"
-            // ["cat2"] +-+> "ed1.eu12p1"
-            // ["cat3"] | +--> "ed1.eu12p1.eu2p31"
-            // [      ] | +--> "ed1.eu12p1.eu2p32"
-            // [      ] +-+> "ed1.eu12p2"
-            // ["cat3"] | +--> "ed1.eu12p2.eu2p31"
-            // [      ] | +--> "ed1.eu12p2.eu2p32"
+            this.eu2.Category.Add(this.cat2);
+            this.ed1.ContainedElement.Add(this.eu2);
         }
 
         private void AddParameter(
@@ -330,42 +235,6 @@ namespace CDP4Composition.Tests.Reporting
         }
 
         [Test]
-        public void VerifyStructure()
-        {
-            var hierarchy = new CategoryHierarchy
-                    .Builder(this.iteration, this.cat1.ShortName)
-                .AddLevel(this.cat2.ShortName)
-                .AddLevel(this.cat3.ShortName)
-                .Build();
-
-            var dataSource = new ReportingDataSourceClass<Row>(
-                this.iteration,
-                hierarchy);
-
-            var rows = dataSource.GetTabularRepresentation();
-
-            // tabular representation is built and category hierarchy is considered
-            Assert.AreEqual(6, rows.Count);
-            Assert.AreEqual("ed1", rows[0].ElementName);
-            Assert.AreEqual("ed1.eu12n1", rows[1].ElementName);
-            Assert.AreEqual("ed1.eu12p1", rows[2].ElementName);
-            Assert.AreEqual("ed1.eu12p1.eu2p31", rows[3].ElementName);
-            Assert.AreEqual("ed1.eu12p2", rows[4].ElementName);
-            Assert.AreEqual("ed1.eu12p2.eu2p31", rows[5].ElementName);
-
-            // skips levels (visibility)
-            var filteredRow = rows[4];
-            Assert.IsFalse(filteredRow.IsVisible);
-
-            // filtered rows have no parameters
-            Assert.IsNull(filteredRow.parameter1);
-            Assert.IsNull(filteredRow.parameter2);
-
-            // prunes unneeded subtrees
-            Assert.IsNull(rows.FirstOrDefault(x => x.ElementName == "ed1.eu12n2"));
-        }
-
-        [Test]
         public void VerifyThatNodeIdentifiesParameters()
         {
             var hierarchy = new CategoryHierarchy
@@ -411,36 +280,28 @@ namespace CDP4Composition.Tests.Reporting
                 hierarchy);
 
             var parameter1 = node.GetColumn<TestParameter1>();
-            Assert.AreEqual(
-                "11",
-                parameter1.GetValue());
+            Assert.AreEqual("11", parameter1.GetValue());
 
             var parameter2 = node.GetColumn<TestParameter2>();
-            Assert.AreEqual(
-                "12",
-                parameter2.GetValue());
+            Assert.AreEqual("12", parameter2.GetValue());
         }
 
         [Test]
         public void VerifyParameterOverrideValueInitialization()
         {
             var hierarchy = new CategoryHierarchy
-                    .Builder(this.iteration, this.cat3.ShortName)
+                    .Builder(this.iteration, this.cat2.ShortName)
                 .Build();
 
             var node = new ReportingDataSourceNode<Row>(
-                this.eu2p31,
+                this.eu1,
                 hierarchy);
 
             var parameter1 = node.GetColumn<TestParameter1>();
-            Assert.AreEqual(
-                "231",
-                parameter1.GetValue());
+            Assert.AreEqual("121", parameter1.GetValue());
 
             var parameter2 = node.GetColumn<TestParameter2>();
-            Assert.AreEqual(
-                "232",
-                parameter2.GetValue());
+            Assert.AreEqual("122", parameter2.GetValue());
         }
 
         [Test]
@@ -482,13 +343,13 @@ namespace CDP4Composition.Tests.Reporting
 
             var children1 = parameter.GetChildren<TestParameter1>().ToList();
             Assert.AreEqual(2, children1.Count);
-            Assert.AreEqual("-21", children1[0].GetValue());
-            Assert.AreEqual("21", children1[1].GetValue());
+            Assert.AreEqual("121", children1[0].GetValue());
+            Assert.AreEqual("221", children1[1].GetValue());
 
             var children2 = parameter.GetChildren<TestParameter2>().ToList();
             Assert.AreEqual(2, children2.Count);
-            Assert.AreEqual("-22", children2[0].GetValue());
-            Assert.AreEqual("22", children2[1].GetValue());
+            Assert.AreEqual("122", children2[0].GetValue());
+            Assert.AreEqual("222", children2[1].GetValue());
         }
     }
 }
